chore(biayaRiilService): solve the test case

+++ api/src/biaya-riil/biaya-riil.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { AppUtil } from 'src/chaincode-service/appUtil.service';
 import { HlfConfig } from 'src/chaincode-service/hlfConfig';
 import { BiayaRiilDTO } from './biaya-riil.dto';
@@ -26,6 +26,8 @@ export class BiayaRiilService {
   }
 
   async findOne(key: string) {
+    if (!key) throw new BadRequestException('Key argument cannot be empty');
+
     const result = await this.hlfConfig.contract.evaluateTransaction(
       'getByKey',
       key,
@@ -35,6 +37,8 @@ export class BiayaRiilService {
   }
 
   async update(key: string, biayariilDTO: BiayaRiilDTO) {
+    if (!key) throw new BadRequestException('Key argument cannot be empty');
+
     const result = await this.hlfConfig.contract.submitTransaction(
       'updateByKey',
       key,
@@ -45,6 +49,8 @@ export class BiayaRiilService {
   }
 
   async remove(key: string) {
+    if (!key) throw new BadRequestException('Key argument cannot be empty');
+
     const result = await this.hlfConfig.contract.submitTransaction(
       'deleteByKey',
       key,
